requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-ngrok==0.0.25
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mysql-connector-python==8.0.27
oauthlib==3.1.1
protobuf==3.19.3
PyMySQL==1.0.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.29
tweepy==4.4.0
urllib3==1.26.8
Werkzeug==2.0.2
